backend/requirements.txt
asgiref==3.8.1
bcrypt==4.0.0
blinker==1.9.0
click==8.3.0
cx_Oracle==8.3.0
Django==4.2
django-bcrypt==0.9.2
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==12.1.0
psycopg2-binary==2.9.11
py-bcrypt==0.4
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.2
Werkzeug==3.1.3
